backend/auth_service/models.py
from typing import Optional

from pydantic import BaseModel, EmailStr, Field


class UserBase(BaseModel):
    email: EmailStr
    full_name: Optional[str] = None


class UserCreate(UserBase):
    password: str = Field(min_length=8)


class UserDB(UserBase):
    id: str | None = None  # Mongo _id as string
    hashed_password: str
    role: str = "pharmacist"


class Token(BaseModel):
    access_token: str
    token_type: str = "bearer"


class TokenPayload(BaseModel):
    sub: str  # subject = user id
    exp: int
