backend/auth_service/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
python-jose>=3.3
passlib[bcrypt]>=1.7
pymongo>=4.6
email-validator>=2.0
python-multipart>=0.0.20
prometheus_client>=0.20
